feat: apply cost data-type and separate source form component
